requirements.txt
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
hydra-core==1.3.2
scipy==1.11.2
jupyter==1.0.0
tqdm==4.66.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
yt-dlp==2023.7.6

